Change code style for the standard java convention.
Pull request changes:
- File names to upper camel case.
- Javadoc style for documentation.
- Create a Main class to execute the other points.
