+++ src/Main.java
@@ -1,0 +1,14 @@
+import points.*;
+
+public class Main {
+
+    /**
+     * Execute points from Main
+     * @param args: command line arguments.
+     */
+    public static void main(String[] args) {
+        // Punto 1
+        Point1.main();
+        Point2.main();
+    }
+}
