+++ src/points/Point3.java
@@ -1,13 +1,15 @@
-/*Escriba una aplicación que reciba tres valores decimales del usuario y muestre la suma, promedio, producto, cociente
-y modulo. Los resultados se deben imprimir en formato decimal con dos cifras significativas.
-*/
+/** Escriba una aplicación que reciba tres valores decimales del
+ * usuario y muestre la suma, promedio, producto, cociente
+ * y modulo. Los resultados se deben imprimir en formato decimal
+ * con dos cifras significativas.
+ */
 
 package points;
 
 import java.util.Scanner;
 
-public class point3 {
-    public static void main(String[] args) {
+public class Point3 {
+    public static void main() {
         Scanner entry = new Scanner(System.in);
 
         System.out.println("Digite el primer decimal: ");
@@ -18,7 +20,6 @@ public class point3 {
 
         System.out.println("Digite el tercer decimal: ");
         float num3 = entry.nextFloat();
-
 
     }
 }
